ai-service/requirements.txt
fastapi
uvicorn
openai-whisper
transformers
torch
python-multipart
pymongo
python-dotenv
dateparser
spacy==3.7.4
numpy==1.26.4